add core and shared module
