+++ src/app/@core/@core.module.ts
@@ -1,15 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { JoinFormComponent } from './join-form/join-form.component';
 import { MainPageComponent } from './main-page/main-page.component';
 import { PostPageComponent } from './post-page/post-page.component';
 import { SharedModule } from '../@shared/@shared.module';
 import { LoginMenuComponent } from './login-menu/login-menu.component';
 
+
 @NgModule({
   declarations: [
-    JoinFormComponent,
     MainPageComponent,
     PostPageComponent,
     LoginMenuComponent
@@ -19,7 +18,6 @@ import { LoginMenuComponent } from './login-menu/login-menu.component';
     SharedModule
   ],
   exports: [
-    JoinFormComponent,
     MainPageComponent,
     PostPageComponent,
     LoginMenuComponent
