+++ src/app/@shared/@shared.module.ts
@@ -6,6 +6,7 @@ import { HomeButtonComponent } from './home-button/home-button.component';
 import { NavigationMenuComponent } from './navigation-menu/navigation-menu.component';
 import { FooterComponent } from './footer/footer.component';
 import { InternshipCardComponent } from './internship-card/internship-card.component';
+import { JoinFormComponent } from './join-form/join-form.component';
 import { JoinInputDateComponent } from './join-input-date/join-input-date.component';
 import { JoinSelectComponent } from './join-select/join-select.component';
 import { JoinSelectHoursComponent } from './join-select-hours/join-select-hours.component';
@@ -28,6 +29,7 @@ export function HttpLoaderFactory(http: HttpClient) {
 
 @NgModule({
   declarations: [
+    JoinFormComponent,
     JoinInputDateComponent,
     JoinSelectComponent,
     JoinInputComponent,
@@ -58,6 +60,7 @@ export function HttpLoaderFactory(http: HttpClient) {
         })
   ],
   exports: [
+    JoinFormComponent,
     JoinInputDateComponent,
     JoinSelectComponent,
     JoinInputComponent,
