+++ src/app/@shared/join-form/join-form.component.scss
