+++ src/app/@shared/join-input-date/join-input-date.component.ts
@@ -4,7 +4,7 @@ import { Component, Input, OnInit } from '@angular/core';
   selector: 'app-join-input-date',
   template:  `<div>
                 <select name="day" class='form-control'>
-                  <option name="day" value="" hidden>{{'POST_PAIGE.JOIN_FORM.InterviewDay' | translate}}</option>
+                  <option name="day" value="" hidden>{{'JOIN_FORM.InterviewDay' | translate}}</option>
                   <option name="day" *ngFor="let day of InterviewDays"
                         value="{day}"
                   >{{day}}
