+++ src/app/@shared/join-select-hours/join-select-hours.component.ts
@@ -4,7 +4,7 @@ import { Component, OnInit, Input } from '@angular/core';
   selector: 'app-join-select-hours',
   template: `<div>
               <select name="hours">
-                <option name="hours" value="" hidden>{{'POST_PAIGE.JOIN_FORM.InterviewHours' | translate}}</option>
+                <option name="hours" value="" hidden>{{'JOIN_FORM.InterviewHours' | translate}}</option>
                 <option name="hours" *ngFor="let hours of InterviewHours"
                       value="{{hours}}"
                 >{{hours}}
