+++ src/assets/i18n/en.json
@@ -1,32 +1,6 @@
 {
     "POST_PAIGE":{
-        "HOME_BUTTON":"Home page",
-        "JOIN_FORM": {
-            "Title":"Join us",
-            "FirstName":"First Name...",
-            "LastName":"Last Name...",
-            "Telephone":"Telephone...",
-            "Location":"Location...",
-            "Email":"Email...",
-            "Skype":"Skype...",
-            "GitHub":"GitHub link...",
-            "English":"English level...",
-            "CV": "Drop o select your CV",
-            "Interview": "Choose suitable time for you for an interview:",
-            "InterviewDay": "Part of the week",
-            "InterviewDays": {
-                "Monday":"Monday", 
-                "Tuesday":"Tuesday", 
-                "Wednesday":"Wednesday", 
-                "Thursday":"Thursday",
-                "Friday":"Friday",
-                "Saturday":"Saturday"
-            },
-            "InterviewHours": "Part of the day",
-            "Agreement": "I agree to the processing and storage of my personal data",
-            "Notification": "I want to receive notifications for similar events",
-            "SubmitBtn": "Submit"
-        }
+        "HOME_BUTTON":"Home page"
     },
     "MAIN_PAIGE":{
         "FILTERS":{
@@ -35,5 +9,33 @@
             "Event_type":"Event type",
             "Internship_status":"Internship status"
         }
+    },
+    "JOIN_FORM": {
+        "Title":"Join us",
+        "FirstName":"First Name...",
+        "LastName":"Last Name...",
+        "Telephone":"Telephone...",
+        "Location":"Location...",
+        "Email":"Email...",
+        "Skype":"Skype...",
+        "GitHub":"GitHub link...",
+        "English":"English level...",
+        "Drop": "Drop or ",
+        "Select":"Select ",
+        "CV": "your CV",
+        "Interview": "Choose suitable time for you for an interview:",
+        "InterviewDay": "Part of the week",
+        "InterviewDays": {
+            "Monday":"Monday", 
+            "Tuesday":"Tuesday", 
+            "Wednesday":"Wednesday", 
+            "Thursday":"Thursday",
+            "Friday":"Friday",
+            "Saturday":"Saturday"
+        },
+        "InterviewHours": "Part of the day",
+        "Agreement": "I agree to the processing and storage of my personal data",
+        "Notification": "I want to receive notifications for similar events",
+        "SubmitBtn": "Submit"
     }
 }
